+++ EatPic-iOS/Sources/Screens/Home/AddMealView.swift
@@ -1,0 +1,29 @@
+////
+////  MealAddView.swift
+////  EatPic-iOS
+////
+////  Created by 이은정 on 7/23/25.
+////
+//
+//import SwiftUI
+//
+//struct AddMealView1: View {
+//    var body: some View {
+//        VStack {
+//
+//            // 사진 추가 ~ 이미지
+//            ZStack {
+//                RoundedRectangle(cornerRadius: 10)
+//                    .fill(Color.gray020)
+//                    .frame(width: 76, height: 76)
+//                
+//                Image("Home/btn_home_add")
+//                
+//            }
+//        }
+//    }
+//}
+//
+//#Preview {
+//    AddMealView()
+//}
